chore: fix tests
On-behalf-of: SAP <matthias.bruns@sap.com>
