+++ bindings/go/cli/integration/add_component_version_integration_test.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"net"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -13,8 +14,13 @@ import (
 	"testing"
 	"time"
 
+	awsconfig "github.com/aws/aws-sdk-go-v2/config"
+	awscreds "github.com/aws/aws-sdk-go-v2/credentials"
+	awss3 "github.com/aws/aws-sdk-go-v2/service/s3"
 	godigest "github.com/opencontainers/go-digest"
 	"github.com/stretchr/testify/require"
+	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/modules/minio"
 
 	"ocm.software/open-component-model/bindings/go/blob/filesystem"
 	"ocm.software/open-component-model/bindings/go/cli/cmd"
@@ -1107,6 +1113,239 @@ components:
 		r.Equal(internal.GitHubRef, stored["ref"],
 			"the ref stays informational next to the pinned commit")
 	})
+}
+
+// Test_Integration_AddComponentVersion_S3Access verifies that a resource declaring an S3
+// access directly in the constructor is recorded as-is, that the s3 digest processor hashes
+// the object, and that the access is then resolved and downloaded via the s3 resource
+// repository with typed S3Credentials/v1 from the ocm config.
+func Test_Integration_AddComponentVersion_S3Access(t *testing.T) {
+	t.Parallel()
+	r := require.New(t)
+	ctx := t.Context()
+
+	const bucket, objectKey = "access-bucket", "path/to/artifact.txt"
+	content := []byte("hello from s3 access")
+
+	registry, err := internal.CreateOCIRegistry(t)
+	r.NoError(err)
+	endpoint, cfgPath := startS3WithObject(t, registry, bucket, objectKey, content)
+
+	const componentName = "ocm.software/s3-access-component"
+	const componentVersion = "v1.0.0"
+
+	// The resource declares an s3 access directly, not an input method.
+	constructorContent := fmt.Sprintf(`
+components:
+- name: %s
+  version: %s
+  provider:
+    name: ocm.software
+  resources:
+  - name: remote-blob
+    version: v1.0.0
+    type: blob
+    relation: external
+    access:
+      type: S3/v2
+      endpoint: %s
+      usePathStyle: true
+      region: us-east-1
+      bucketName: %s
+      objectKey: %s
+      mediaType: text/plain
+`, componentName, componentVersion, endpoint, bucket, objectKey)
+	constructorPath := filepath.Join(t.TempDir(), "constructor.yaml")
+	r.NoError(os.WriteFile(constructorPath, []byte(constructorContent), os.ModePerm))
+
+	addCMD := cmd.New()
+	addCMD.SetArgs([]string{
+		"add",
+		"component-version",
+		"--repository", fmt.Sprintf("http://%s", registry.RegistryAddress),
+		"--constructor", constructorPath,
+		"--config", cfgPath,
+	})
+	addCtx, cancel := context.WithTimeout(ctx, 60*time.Second)
+	defer cancel()
+	r.NoError(addCMD.ExecuteContext(addCtx), "add cv with s3 access should succeed")
+
+	// The access is recorded as-is; the resource keeps its s3 access spec.
+	repo := registry.Connect(t)
+	desc, err := repo.GetComponentVersion(ctx, componentName, componentVersion)
+	r.NoError(err)
+	r.Len(desc.Component.Resources, 1)
+	res := desc.Component.Resources[0]
+	r.Equal("remote-blob", res.Name)
+	r.NotNil(res.Access, "resource should carry an s3 access spec")
+	r.Equal("S3/v2", res.Access.GetType().String())
+
+	// The s3 digest processor fetched the object once at add time and set the digest.
+	r.NotNil(res.Digest, "s3 digest processor should have set a digest")
+	r.Equal("SHA-256", res.Digest.HashAlgorithm)
+	r.Equal("genericBlobDigest/v1", res.Digest.NormalisationAlgorithm)
+	r.Equal(godigest.FromBytes(content).Encoded(), res.Digest.Value,
+		"digest value must be the SHA-256 of the object")
+
+	// download resource resolves the s3 access via the registered s3 resource repository.
+	output := filepath.Join(t.TempDir(), "downloaded")
+	downloadCMD := cmd.New()
+	downloadCMD.SetArgs([]string{
+		"download",
+		"resource",
+		fmt.Sprintf("http://%s//%s:%s", registry.RegistryAddress, componentName, componentVersion),
+		"--identity", "name=remote-blob,version=v1.0.0",
+		"--output", output,
+		"--config", cfgPath,
+	})
+	r.NoError(downloadCMD.ExecuteContext(ctx), "download resource should resolve the s3 access")
+
+	outputBlob, err := filesystem.GetBlobFromOSPath(output)
+	r.NoError(err)
+	r.Equal(content, readAllFromBlob(t, outputBlob), "downloaded content should match the object in the bucket")
+
+	// get component-version reads the s3-access component back through the CLI.
+	getOutput := new(bytes.Buffer)
+	getCMD := cmd.New()
+	getCMD.SetOut(getOutput)
+	getCMD.SetArgs([]string{
+		"get",
+		"component-version",
+		fmt.Sprintf("http://%s//%s:%s", registry.RegistryAddress, componentName, componentVersion),
+		"--config", cfgPath,
+		"--output", "json",
+	})
+	r.NoError(getCMD.ExecuteContext(ctx), "get component-version should succeed for the s3-access component")
+
+	out := getOutput.String()
+	r.Contains(out, componentName, "output should contain the component name")
+	r.Contains(out, "remote-blob", "output should contain the resource")
+	r.Contains(out, "S3/v2", "output should contain the s3 access type")
+}
+
+// Test_Integration_S3Input verifies that an S3 input reads the object from the bucket with
+// typed S3Credentials/v1 from the ocm config and stores it as a local blob.
+func Test_Integration_S3Input(t *testing.T) {
+	t.Parallel()
+	r := require.New(t)
+	ctx := t.Context()
+
+	const bucket, objectKey = "input-bucket", "path/to/artifact.txt"
+	content := []byte("hello from s3 input")
+
+	registry, err := internal.CreateOCIRegistry(t)
+	r.NoError(err)
+	endpoint, cfgPath := startS3WithObject(t, registry, bucket, objectKey, content)
+
+	const componentName = "ocm.software/s3-input-component"
+	const componentVersion = "v1.0.0"
+
+	constructorContent := fmt.Sprintf(`
+components:
+- name: %s
+  version: %s
+  provider:
+    name: ocm.software
+  resources:
+  - name: remote-blob
+    version: v1.0.0
+    type: blob
+    input:
+      type: S3/v2
+      endpoint: %s
+      usePathStyle: true
+      region: us-east-1
+      bucketName: %s
+      objectKey: %s
+      mediaType: text/plain
+`, componentName, componentVersion, endpoint, bucket, objectKey)
+	constructorPath := filepath.Join(t.TempDir(), "constructor.yaml")
+	r.NoError(os.WriteFile(constructorPath, []byte(constructorContent), os.ModePerm))
+
+	// add: the s3 input reads the object and stores it as a local blob.
+	addCMD := cmd.New()
+	addCMD.SetArgs([]string{
+		"add",
+		"component-version",
+		"--repository", fmt.Sprintf("http://%s", registry.RegistryAddress),
+		"--constructor", constructorPath,
+		"--config", cfgPath,
+	})
+	addCtx, cancel := context.WithTimeout(ctx, 60*time.Second)
+	defer cancel()
+	r.NoError(addCMD.ExecuteContext(addCtx), "add cv with s3 input should succeed")
+
+	repo := registry.Connect(t)
+	desc, err := repo.GetComponentVersion(ctx, componentName, componentVersion)
+	r.NoError(err)
+	r.Len(desc.Component.Resources, 1)
+	res := desc.Component.Resources[0]
+	r.Equal("remote-blob", res.Name)
+
+	blobData, _, err := repo.GetLocalResource(ctx, componentName, componentVersion, res.ToIdentity())
+	r.NoError(err)
+	r.Equal(content, readAllFromBlob(t, blobData), "local blob should hold the object from the bucket")
+}
+
+// startS3WithObject starts MinIO holding content at bucket/key and writes an ocmconfig
+// with typed S3Credentials/v1 for it next to the registry credentials. It returns the
+// S3 endpoint and the config path.
+func startS3WithObject(t *testing.T, registry *internal.OCIRegistry, bucket, key string, content []byte) (endpoint, cfgPath string) {
+	t.Helper()
+	r := require.New(t)
+	ctx := t.Context()
+
+	container, err := minio.Run(ctx, "minio/minio")
+	r.NoError(err)
+	t.Cleanup(func() { r.NoError(testcontainers.TerminateContainer(container)) })
+	hostPort, err := container.ConnectionString(ctx)
+	r.NoError(err)
+	host, port, err := net.SplitHostPort(hostPort)
+	r.NoError(err)
+	endpoint = "http://" + hostPort
+
+	awsCfg, err := awsconfig.LoadDefaultConfig(ctx,
+		awsconfig.WithRegion("us-east-1"),
+		awsconfig.WithCredentialsProvider(awscreds.NewStaticCredentialsProvider(container.Username, container.Password, "")),
+	)
+	r.NoError(err)
+	client := awss3.NewFromConfig(awsCfg, func(o *awss3.Options) {
+		o.BaseEndpoint = &endpoint
+		o.UsePathStyle = true
+	})
+	_, err = client.CreateBucket(ctx, &awss3.CreateBucketInput{Bucket: &bucket})
+	r.NoError(err)
+	_, err = client.PutObject(ctx, &awss3.PutObjectInput{Bucket: &bucket, Key: &key, Body: bytes.NewReader(content)})
+	r.NoError(err)
+
+	cfg := fmt.Sprintf(`
+type: generic.config.ocm.software/v1
+configurations:
+- type: credentials.config.ocm.software
+  consumers:
+  - identity:
+      type: OCIRegistry
+      hostname: %[1]q
+      port: %[2]q
+      scheme: http
+    credentials:
+    - type: Credentials/v1
+      properties:
+        username: %[3]q
+        password: %[4]q
+  - identity:
+      type: S3
+      hostname: %[5]q
+      port: %[6]q
+      scheme: http
+    credentials:
+    - type: S3Credentials/v1
+      accessKeyId: %[7]q
+      secretAccessKey: %[8]q
+`, registry.Host, registry.Port, registry.User, registry.Password, host, port, container.Username, container.Password)
+	cfgPath = filepath.Join(t.TempDir(), "ocmconfig.yaml")
+	r.NoError(os.WriteFile(cfgPath, []byte(cfg), os.ModePerm))
+	return endpoint, cfgPath
 }
 
 func Test_Integration_AddComponentVersion_SpecificationValidation(t *testing.T) {
